chore(token-definitions): move token definitions to it's package
