+++ suite-common/token-definitions/package.json
@@ -13,7 +13,8 @@
         "test:unit": "yarn g:jest -c=../../jest.config.base.js"
     },
     "dependencies": {
-        "@suite-common/wallet-config": "workspace:*"
+        "@suite-common/wallet-config": "workspace:*",
+        "@suite-common/wallet-types": "workspace:*"
     },
     "devDependencies": {
         "@types/fs-extra": "^11.0.4",
