+++ suite-common/token-definitions/src/types.ts
@@ -1,3 +1,5 @@
+import { NetworkSymbol } from '@suite-common/wallet-config';
+
 export type SimpleTokenStructure = string[];
 
 export interface AdvancedTokenStructure {
@@ -10,3 +12,20 @@ export enum DefinitionType {
     NFT = 'nft',
     COIN = 'coin',
 }
+
+export type TokenDefinitionsState = {
+    [key in NetworkSymbol]?: TokenDefinitions;
+};
+
+export type TokenDefinitionsRootState = { wallet: { tokenDefinitions: TokenDefinitionsState } };
+
+type TokenDefinition = {
+    error: boolean;
+    data?: SimpleTokenStructure;
+    isLoading: boolean;
+};
+
+export type TokenDefinitions = {
+    [DefinitionType.COIN]?: TokenDefinition;
+    [DefinitionType.NFT]?: TokenDefinition;
+};
