+++ suite-common/wallet-core/src/index.ts
@@ -33,7 +33,6 @@ export * from './token-definitions/tokenDefinitionsSelectors';
 export * from './token-definitions/tokenDefinitionsReducer';
 export * from './token-definitions/tokenDefinitionsThunks';
 export * from './token-definitions/tokenDefinitionsMiddleware';
-export * from './token-definitions/tokenDefinitionsTypes';
 export * from './stake/stakeActions';
 export * from './stake/stakeReducer';
 export * from './stake/stakeSelectors';
