+++ suite-common/wallet-types/package.json
@@ -13,7 +13,6 @@
         "@suite-common/intl-types": "workspace:*",
         "@suite-common/metadata-types": "workspace:*",
         "@suite-common/suite-config": "workspace:*",
-        "@suite-common/token-definitions": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
         "@suite-common/wallet-constants": "workspace:*",
         "@trezor/blockchain-link-types": "workspace:*",
