--- suite-common/wallet-types/src/tokenDefinitions.ts
@@ -1,12 +1,0 @@
-import { DefinitionType, SimpleTokenStructure } from '@suite-common/token-definitions';
-
-type TokenDefinition = {
-    error: boolean;
-    data?: SimpleTokenStructure;
-    isLoading: boolean;
-};
-
-export type TokenDefinitions = {
-    [DefinitionType.COIN]?: TokenDefinition;
-    [DefinitionType.NFT]?: TokenDefinition;
-};
