+++ suite-native/discovery/tsconfig.json
@@ -9,6 +9,9 @@
             "path": "../../suite-common/suite-types"
         },
         {
+            "path": "../../suite-common/token-definitions"
+        },
+        {
             "path": "../../suite-common/wallet-config"
         },
         {
