+++ suite-native/module-accounts-import/tsconfig.json
@@ -10,6 +10,9 @@
             "path": "../../suite-common/redux-utils"
         },
         {
+            "path": "../../suite-common/token-definitions"
+        },
+        {
             "path": "../../suite-common/validators"
         },
         {
