+++ suite-native/transactions/src/components/TransactionDetail/TransactionDetailData.tsx
@@ -13,7 +13,7 @@ import {
 import { EthereumTokenTransfer, WalletAccountTransaction } from '@suite-native/ethereum-tokens';
 import { Translation } from '@suite-native/intl';
 import { Link } from '@suite-native/link';
-import { TokenDefinitionsRootState } from '@suite-common/wallet-core';
+import { TokenDefinitionsRootState } from '@suite-common/token-definitions';
 
 import { TransactionDetailSummary } from './TransactionDetailSummary';
 import { TransactionDetailRow } from './TransactionDetailRow';
