+++ suite-native/transactions/tsconfig.json
@@ -16,6 +16,9 @@
             "path": "../../suite-common/suite-utils"
         },
         {
+            "path": "../../suite-common/token-definitions"
+        },
+        {
             "path": "../../suite-common/wallet-config"
         },
         {
